extract route component
