+++ extension/src/panel/app-routes/list-routes/ListRoutes.tsx
@@ -1,52 +1,26 @@
-import {
-  Box,
-  BoxButton,
-  Button,
-  ConnectionStack,
-  Divider,
-  Flex,
-  useConfirmationModal,
-} from '@/components'
-import { ZodiacRoute } from '@/types'
-import {
-  useSelectedRouteId,
-  useZodiacRoute,
-  useZodiacRoutes,
-} from '@/zodiac-routes'
-import { formatDistanceToNow } from 'date-fns'
+import { Button, Divider } from '@/components'
+import { useSelectedRouteId, useZodiacRoutes } from '@/zodiac-routes'
 import { nanoid } from 'nanoid'
-import React from 'react'
 import { useNavigate } from 'react-router-dom'
-import { useClearTransactions } from '../../state/transactionHooks'
-import { asLegacyConnection } from '../legacyConnectionMigrations'
-import { ConnectedIcon } from './ConnectedIcon'
-import { DisconnectedIcon } from './DisconnectedIcon'
-import classes from './style.module.css'
+import { Route } from './Route'
 
 export const ListRoutes = () => {
   const [, selectRoute] = useSelectedRouteId()
   const routes = useZodiacRoutes()
   const navigate = useNavigate()
 
-  const handleLaunch = (routeId: string) => {
-    selectRoute(routeId)
-    navigate('/')
-  }
-  const handleModify = (routeId: string) => {
-    navigate('/routes/' + routeId)
-  }
-
-  const handleCreate = () => {
-    const newRouteId = nanoid()
-    navigate('/routes/' + newRouteId)
-  }
-
   return (
     <div className="flex flex-1 flex-col gap-4 px-6 py-8">
       <div className="flex items-center justify-between">
         <h2 className="text-2xl">Pilot Routes</h2>
 
-        <Button onClick={handleCreate} className={classes.addConnection}>
+        <Button
+          onClick={() => {
+            const newRouteId = nanoid()
+            navigate('/routes/' + newRouteId)
+          }}
+          className="px-6 py-1"
+        >
           Add Route
         </Button>
       </div>
@@ -54,137 +28,19 @@ export const ListRoutes = () => {
       <Divider />
 
       {routes.map((route) => (
-        <RouteItem
+        <Route
           key={route.id}
           route={route}
-          onLaunch={handleLaunch}
-          onModify={handleModify}
+          onLaunch={(routeId) => {
+            selectRoute(routeId)
+
+            navigate('/')
+          }}
+          onModify={(routeId) => {
+            navigate('/routes/' + routeId)
+          }}
         />
       ))}
     </div>
-  )
-}
-
-interface RouteItemProps {
-  route: ZodiacRoute
-  onLaunch: (routeId: string) => void
-  onModify: (routeId: string) => void
-}
-
-const RouteItem: React.FC<RouteItemProps> = ({ onLaunch, onModify, route }) => {
-  const { connected, connect, chainId } = useZodiacRoute(route.id)
-  const { route: currentlySelectedRoute } = useZodiacRoute()
-  const [getConfirmation, ConfirmationModal] = useConfirmationModal()
-  const { hasTransactions, clearTransactions } = useClearTransactions()
-
-  const confirmClearTransactions = async () => {
-    if (!hasTransactions) {
-      return true
-    }
-
-    const confirmation = await getConfirmation(
-      'Switching the Piloted Safe will empty your current transaction bundle.'
-    )
-
-    if (!confirmation) {
-      return false
-    }
-
-    clearTransactions()
-
-    return true
-  }
-
-  const handleModify = () => onModify(route.id)
-
-  const handleLaunch = async () => {
-    // we continue working with the same avatar, so don't have to clear the recorded transaction
-    const keepTransactionBundle =
-      currentlySelectedRoute && currentlySelectedRoute.avatar === route.avatar
-
-    const confirmed =
-      keepTransactionBundle || (await confirmClearTransactions())
-
-    if (!confirmed) {
-      return
-    }
-
-    if (connected) {
-      onLaunch(route.id)
-      return
-    }
-
-    if (!connected && connect) {
-      const success = await connect()
-      if (success) {
-        onLaunch(route.id)
-        return
-      }
-    }
-
-    handleModify()
-  }
-
-  return (
-    <>
-      <div className="relative">
-        <BoxButton
-          className={classes.connectionItemContainer}
-          onClick={handleLaunch}
-        >
-          <Flex direction="column" gap={4}>
-            <Flex
-              direction="row"
-              gap={2}
-              justifyContent="space-between"
-              className={classes.labelContainer}
-            >
-              <Flex direction="row" alignItems="center" gap={3}>
-                <Box className={classes.connectionIcon}>
-                  {connected ? (
-                    <ConnectedIcon>Pilot wallet is connected</ConnectedIcon>
-                  ) : connect ? (
-                    <ConnectedIcon color="orange">
-                      Pilot wallet is connected to a different chain
-                    </ConnectedIcon>
-                  ) : (
-                    <DisconnectedIcon>
-                      Pilot wallet is not connected
-                    </DisconnectedIcon>
-                  )}
-                </Box>
-                <h2>
-                  {route.label || <em>Unnamed route</em>}
-
-                  <div className="flex items-end gap-2 text-sm font-normal">
-                    <div className="text-xs uppercase text-zodiac-light-mustard">
-                      Last Used
-                    </div>
-                    <div className="font-mono opacity-70">
-                      {route.lastUsed ? (
-                        `${formatDistanceToNow(route.lastUsed)} ago`
-                      ) : (
-                        <>N/A</>
-                      )}
-                    </div>
-                  </div>
-                </h2>
-              </Flex>
-            </Flex>
-
-            <ConnectionStack
-              chainId={chainId}
-              connection={asLegacyConnection(route)}
-              addressBoxClass={classes.addressBox}
-              className={classes.connectionStack}
-            />
-          </Flex>
-        </BoxButton>
-        <BoxButton onClick={handleModify} className={classes.modifyButton}>
-          Modify
-        </BoxButton>
-      </div>
-      <ConfirmationModal />
-    </>
   )
 }
