+++ extension/src/panel/app-routes/list-routes/style.module.css
@@ -1,33 +1,8 @@
-h2 {
-  font-size: 1.5em;
-  margin: 0em;
-}
-
-button.removeButton {
-  height: auto;
-  width: auto;
-  aspect-ratio: 1/1;
-  padding: 0 8px;
-  border: 1px solid crimson;
-}
-
-button.removeButton:hover {
-  background-color: rgba(220, 20, 60, 0.2);
-}
-
-.addConnection {
-  width: auto;
-  padding: 4px 24px;
-}
-
 .connection {
   position: relative;
 }
 
 .connectionItemContainer {
-  width: 100%;
-  background-color: rgba(39, 38, 30, 0.7);
-  border-color: rgba(255, 255, 255, 0.3);
   box-shadow: 0px 2px 10px 0px rgba(0, 0, 0, 0.1);
   padding: var(--spacing-4);
 }
@@ -46,67 +21,8 @@ button.removeButton:hover {
   padding-top: 34px;
 }
 
-.connectionStack {
-  min-width: 505px;
-}
-
-.modifyButton {
-  position: absolute !important;
-  top: var(--spacing-4);
-  right: var(--spacing-4);
-  background: none;
-  padding: var(--spacing-1) var(--spacing-3);
-}
-
-.modifyButton:before {
-  content: none;
-}
-
-.connectionIcon {
-  position: relative;
-  border: 1px solid;
-  padding: 0;
-  border-color: var(--border-color);
-  border-radius: 50px;
-  background: rgb(0 0 0 / 0.3);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 4px;
-}
-.connectionIcon:before {
-  content: ' ';
-  position: absolute;
-  z-index: 1;
-  top: -4px;
-  left: -4px;
-  right: -4px;
-  bottom: -4px;
-  border: 1px solid var(--border-color);
-  pointer-events: none;
-  border-radius: 50px;
-}
-
-.connectionIcon svg {
-  border-radius: 50px;
-}
-
 .infoContainer {
   padding: 0 0 0 45px;
-}
-
-.labelContainer {
-  overflow: hidden;
-}
-
-.labelContainer h2 {
-  font-size: 1.4em;
-  margin-bottom: 0;
-  line-height: 1.2;
-  text-align: left;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
 }
 
 .infoDatum {
