+++ extension/tailwind.config.js
@@ -18,6 +18,8 @@ export default {
           'dark-green': 'rgba(0, 166, 56, 0.47)',
 
           'light-red': 'rgba(220, 20, 60)',
+
+          'very-dark-blue': 'rgba(39, 38, 30)',
         },
       },
     },
